+++ src/components/Scrollable/types.ts
@@ -1,0 +1,6 @@
+import * as PropTypes from 'prop-types';
+import {ValidationMap} from 'react';
+
+export const contextTypes: ValidationMap<any> = {
+  scrollToPosition: PropTypes.func,
+};